# MyProject/bookstore-front/src/app/service/model/book.ts
/**
 * BookStore APIs
 * BookStore APIs exposed from a Java EE back-end to an Angular front-end
 *
 * OpenAPI spec version: 1.0.0
 * Contact: antonio.goncalves@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

/**
 * Book resource representation
 */
export interface Book {
    /**
     * Identifier
     */
    id?: number;
    /**
     * Title of the book
     */
    title: string;
    /**
     * Description of the book
     */
    description?: string;
    unitCost?: number;
    isbn?: string;
    publicationDate?: Date;
    nbOfPages?: number;
    imageURL?: string;
    language?: Book.LanguageEnum;
}
export namespace Book {
  export enum LanguageEnum {
    ENGLISH = <any> 'ENGLISH',
    FRENCH = <any> 'FRENCH',
    SPANISH = <any> 'SPANISH',
    PORTUGUESE = <any> 'PORTUGUESE',
    ITALIAN = <any> 'ITALIAN',
    FINISH = <any> 'FINISH',
    GERMAN = <any> 'GERMAN',
    DUTCH = <any> 'DUTCH',
    RUSSIAN = <any> 'RUSSIAN'
  }
}